fix: now you can pass &str or String to api key and url
